pathbar: Change cursor to text when action is loation entry
As part of commit f65c2921b043 ("window: Open location entry when
current path is clicked") we allow entering location entry
by clicking the current folder button.  This, however, is not at
all obvious, not only defeats the original intent of the change
but also leads to issues such as the inability to maximize the
nautilus window (by double clicking the current folder).
Hopefully if we make the action more obvious, errors will be
less likely.

Furthermore, with the pending elimination of Other Places
(and replacement by Network View), Computers would be the lone
remaining item in the sidebar.  Since there is a desire to
drop Computer from the sidebar, we need to make it clear
for users that want access to the root file system of how
to get there