pathbar: Change cursor to text when action is loation entry
As part of commit f65c2921b043 ("window: Open location entry when
current path is clicked") we allow entering location entry
by clicking the current folder button.  This, however, is not at
all obvious, and not only defeats the original intent of the change
but also leads to issues such as the inability to maximize the
nautilus window (by double clicking the current folder).
Hopefully if we make the action more obvious, errors will be
less likely.

Furthermore, with the pending elimination of Other Places
(and replacement by Network View), there's one less way to
access the root file system. This makes the discoverability
of the location entry even more important for people who
want to access '/'.