window-slot: Don't duplicate bookmark in back list
This may happen when opening item location from results of the same
directory, because location changes (from search to directory) but
the history bookmark doesn't (because search is never a bookmark in
history).

So, explicitly prevent prepending the head of the back list to the
list itself a second time, which duplicates the entry unnecessarily
and generates warnings.