gtkplacessidebar: Reorganize places
The XDG directories are always present in the sidebar and not removable
whether or not they are relevant to someone's use case which means the
side bar has several items populating it regardless, pushing relevant
bookmarks further down the list, potentially scrolled out of view.
Also, we need to make space for the upcoming Network place and some
internal storage units which are are coming back to the sidebar.

Let's reduce the number of default sidebar locations by turning all
XDG user directories but Downloads back into regular bookmarks.[0]

Also reposition the Home to the first place, as the default location.
And move the bookmarks and cloud providers sections above the mounts.
This keeps the default bookmarks closer to their original place and
prepares to receive internal partitions (which may overflow the sidebar
and scroll out of view.

Part of: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3012

[0] This assumes a default set of bookmarks including these directories
    is created by xdg-user-dirs-update-gtk on first login. Ensuring it
    is installed, running on startup, and working correctly is a system
    integration and quality assurance task for vendors/administrators.