file-operations: Increment unique filename count
A regression in commit d4526131c1 ("filename-util: add conflict
name creation helper") causes an infinite loop when creating a
file (i.e. from template) where the suffix needs to increment.
This also led to an unexpected behavior of causing the first
duplicate to be suffixed with (1) instead of (2).

Pre-increment the counter to avoid the infinite loop and
start a (2).