window-slot: Add stack widget
For the global search mode we are going to need to cover contents,
including location banners, with a status page.

We currently derive from a box, and GtkStack is final, so instead
derive from AdwBin, have its direct child be a stack, then add a
vertical box to the stack.