query-editor: Ensure clear button is hidden when empty
We have an idle callback to deal with search text changes while
avoiding performance impact. We also freeze this completely if
the text changes are done programatically and not from user
typing.

But the idle callback is also responsible for hiding the clear
icon, which means we don't hide it if the entry is cleared
programatically (e.g. when leaving search using the Escape key
or the toggle button in the headerbar).

Since there is no performance cost to changing this visibility
of the child widget, we can just move this code to the synch
hanlder, before the change_frozen interrupt.


(cherry picked from commit 49c4366aef7c98e5cafb187aeffccb1bcfdbaa10)