files-view: Add gestures to open background menu in empty view
Commit c653f880909 ("list-base: Don't inherit from NautilusFilesView")
broke our ability to right click on an empty view to open the background
context menu because the list-base is no longer a child of files-view.
We could move the gestures from list-base to files-view, but that's
problematic not only because of the weird hack we use (to deny
background click after clicking on a view item, but also because we
might want to not allow background click (for example in the network
view), and claiming the background click in such a view would
interrupt the gesture for the underlying view (i.e. GtkListView).

So let's take the path of least resistance and just add the gesture
on the empty state itself.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3281