window-slot: Keep baner in sync with searched location
When searching in trash and then expanding the search scope to global,
the trash banner sticks around incorrectly.

Let's update the banner on every location change, even ones which do
not change navigation history, and if searching, keep the banner for
the searched location. For global search this location is NULL, which
unsets the banner, as expected.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/3299