placessidebar: Don't add rogue mounts to top section
We add mounts to the sidebar in the following order:

  * First, those belonging to drives; this ensures physical devices
     are prioritized.
  * Second, those belonging to volumes (e.g. GOA accounts)
  * Lastly, other mounts (i.e. unusual stuff)

While the first two are added to the "MOUNTS" section, in the same
order, the last group is added to the "COMPUTER" section (the one
with Recent, Starred, Home, user dirs and Trash).

This has been working like this since the 3.0 sidebar redesign,
all the way back in 2010: a59e586adf32eea64c9c4ae08a5482d2a9c37fb1

That commit refers to a long-gone wiki page, but, regardless, I
have reasons to believe this was not intentional:

 * the COMPUTER section comes before the MOUNTS section, this means
   whe unusual mounts effectively get sorted before the other mounts,
   which runs counter to the effort put into adding them last;
 * it just doesn't make sense to group mounts with the first group,
   which has always been supposed to end with Trash in every mockup
   I've seen in since 2010.
 * given how unusual these rogue mounts are, such a mistake could
   easily have been kept around all these years.
 * the COMPUTER section has translated names which we do not want to
   ellipsize, but mounts can have arbitrarily long names we do want
   to ellipize.

As such, let's just add them to the mounts section, after the others.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3291