list-base: Handle events on empty page
The changes in commit c653f880 ("list-base: Don't inherit
from NautilusFilesView"), broke our gestures on the empty status page
which is still a child of files-view. Rather than attempt to fix this
through files view, as done in commit 95f3db44 ("files-view: Add
gestures to open background menu in empty view"), which fixed the click
gestures, but not the dnd gestures, let's make the more appropriate
change of making the empty status page a child of list-base.

Having the empty status page as a child of the list-base is more logical
because it represents the view, and ultimately our goal is to remove
all view components for files-view (despiste its name).

The empty page is still managed externally, to make it reusable accross
view mode changes. Being an overlay also allows its visibility to be
externally controlled.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3309

(Commit message originally by Corey Berla)