file-operations: Preserve permissions copying from r/o non-isofs mounts
When copying files from readonly mounts we only reset the permissions of
the source if those are of filesystem type isofs. The original behaviour
was meant to be a workaround for isofs optical disks [1][2] for which the
kernel driver sets permissions of all inodes to r-x if the POSIX
permission extension (rock ridge) is not available [1].

As we can't detect the presence of such an extension from gio we can only
guess and reset the permissions for all isofs mounts. This is still
wrong [3], but reduces the scope of the bug to isofs.

[1] https://bugzilla.gnome.org/show_bug.cgi?id=167102
[2] https://bugzilla.gnome.org/show_bug.cgi?id=550816
[3] https://gitlab.gnome.org/GNOME/nautilus/-/issues/2873