file-conflict-dialog: Improve mnemonics of "Replace" & "Rename" buttons
Because of the "Reset" button, the "Replace and "Rename" button
use the letter "P" and "N" respectively for their mnemonics,
which is odd, since the letter "R" would be more practical
to use; and "Replace and "Rename" never appear together.

So, use the letter "R" for "Replace and "Rename" and "E"
for the "Reset" button.