Revert "files-view: Move focus function from list base"
This reverts commit ba665f537102ae022b4746bb3214227f7c6fdb6e.

NautilusFilesView should be agnostic to the inner widgetry. So,
it should not be GtkListItemWidget|listitem.select action.

Also, moving this whole function is not necessary. The true cause of
the bug[0] was that NautilusFilesView was not forwarding the focus
event  its child. Which turned out to be an AdwBin bug. Good news is
this bug has been fixed in libadwaita main[1] and stable[2] branches.

So let's revert this workaround.

[0] https://gitlab.gnome.org/GNOME/nautilus/-/issues/3354
[1] https://gitlab.gnome.org/GNOME/libadwaita/-/commit/0e30d48f6687d89262564b236a8c2cad7c9336c6
[2] https://gitlab.gnome.org/GNOME/libadwaita/-/commit/9962168906acaad873a52670cc1fdc77b67d992c