general: Don't treat root directory as separator
This partially reverts commit c5216c47da5c5792b03df41aadd1190fac5a4823

G_DIR_SEPARATOR_S exists primarily as convenience to make GLib-based
code portable to Microsoft Windows, which is not relevant to this app.
We can still use it anyway when we want to be clear that a slash is
being used as a dir separator.

The root directory's path is "/". While that's the exact same character
as a directory separator, its meaning is completely different.