file-chooser: Introduce new window class
We want to reuse most of our window widgets in the FileChooser dialog,
but we cannot use the actual NautilusWindow class for a few reasons:

  * It's a GtkApplicationWindow subclass, and we don't want FileChooser
    dialogs to be treated as part of the Files app while invoked from
    other apps.
  * FileChooser dialog, as a modal, is expected to behaver differently
    from NautilusWindow is many ways.
  * FileChooser doesn't need a tab bar and tab view, progress indicator,
    nor many other features from NautilusWindow. And the reverse is also
    true, as the FileChooser dialog has an always-on bottom toolbar and
    a different adaptive design for narrow windows.

As such, we introduce a new toplevel window class (NautilusFileChooser)
which is a new composite widget that reuses internal components of the
NautilusWindow class.

This FileChooser window features a sidebar, a top toolbar, a view (as
part of a single NautilusWindowSlot instance), and a bottom toolbar. So,
it looks similar, but not exactly the same, as NautilusWindow.

At this point, this window class is still unused and inert. It's going
to get fuctional and going to be used by NautilusPortal next commit.

More about the code design rationale at
https://gitlab.gnome.org/GNOME/nautilus/-/work_items/3402

Part of https://gitlab.gnome.org/GNOME/nautilus/-/issues/3401