file: Don't rely on application-x-generic
`application-x-generic` is our preferred fallback themed icon. However,
it's  not present in the icon naming specification[0] and, as such, is
not included in the icon themes some downstreams adopt.

Instead of immediately resorting to the unthemed fallback texture, we
can give a second themed alternative which is likely available in any
icon theme: text-x-generic.

[0] https://specifications.freedesktop.org/icon-naming-spec/icon-naming-spec-latest.html

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/3268