window-slot: Inline location loading helper function
The bigger code block of activating a pending file only applies for one
caller, so the helper function did not deduplicate any code and really
only created more overhead, making the code harder to follow.

One significant side affect of this refactoring is that we don't pass
`self->pending_selection` by value to `nautilus_view_set_selection()`.
This resolves a crash in the case when `nautilus_view_set_location()`
will reach a code [1] that uses and frees `self->pending_selection`,
thus leaving the pointer passed by value for selection to become a
dangling pointer if it was pointing to `self->pending_selection`.

[1] https://gitlab.gnome.org/GNOME/nautilus/-/blob/d046bf4a8b1111082c9d3ba6ef1ddf160122a077/src/nautilus-files-view.c#L3896

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3036