list-base: Move long press gesture to child view
When the user tries to select by touch dragging, GtkListBase's drag
gesture is triggered. However, after the delay factor of the long press
has passed, the long press gesture is also triggered and cancels the
user's selection.

This is because both long press and drag gesture event controllers are
set in the bubble phase. The drag threshold is used by both long press
and drag. When this threshold is exceeded, the drag gesture will
claim the sequence and no further updates will be sent to the long press
gesture. This leads to the long press timeout being triggered.

Move the long press gesture to the same widget that has the drag
gesture. To do so, create a new method in NautilusListBase for gestures
that should be attached to the child and not directly to the list base.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2872
Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/3388