files-view: Don't always reload on global search
In [1], global search was excluded from skipping the reloading of
the view like normal search in order for the current content of the
view to not appear in the global search for a short time until the
results were loaded. This causes global search results to flicker
for every key stroke.

Instead, make the condition more specific so that it will not
affect loading between search queries.

[1] ce47e613726ee937c2946be8659367385e2fe7b9

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3403