proprties: Don't scale up the image
In [1], a frame and an overlay were introduced for the redesign of the
thumbnail editor so that the new edit and clear buttons will not be
entirely inside the image frame. This was inside a GtkStack with
another a image which caused it to be scaled since the stack enforces
size homogeneity by default. This is undesirable, so disable it.

[1] 1c70bab1f0c5da26c78fc5530d3256c2f6ba0434

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3422#note_2128815