file-utilities: Provide builtin special dirs info
The built-in special directories are set up in the `user-dirs.dirs`
configuration file. We have to cope with configurations where some
special directories are disabled (i.e. pointing to $HOME) or set
to the same path as other special directories (in which case we
don't want them to be shown twice in the sidebar).

This logic is mixed with the sidebar UI code however, which makes
it hard to reuse for an Edit Sidebar UI.

So, extract it into a helper function. The configuration is loaded
only once, as g_get_user_special_dir() never changes its return
value after the initial loading.