files-view: Make logic order look intentional
The selection check is done after the emission of delayed singals which
may change it, as per commit 09f4546d50fa86a61e3e29a900ff5026303e544a.

However, nowadays we have a habit of combining declaration and
assingment into a single statement. So, during future hacking, one
might be tempted to move the assignments into the declaration lines,
effectively reverting that commit.

Let's keep declaration and assingment together but after that call,
and also add a comment to make it clear this order is intentional.