placessidebar: Update enum symbol
Around the 3.0 release, the sidebar featured section headers and there
was a section whose header was "Computer".

We still carry that name in an enum symbol. It makes no sense nowadays
if you don't know the history. So, let's rename it to the name by
which it is informally called in current design mockups.