window-slot: Import open-location action
The Network view's server address bar relies on this action to open
the destination after the mount is ready.

The upcoming FileChooser is not going to be a NautilusWindow, so we
cannot rely on actions from that class, otherwise the address bar
wouldn't work correctly on FileChooser.

It should be a slot action anyway.

Part of https://gitlab.gnome.org/GNOME/nautilus/-/work_items/3413