window: Erase wrong comment and workaround
The commit[0] which introduced this comment aimed to fix a bug which was
caused by double-clicking a folder on the view.[1]

This change was ineffective as per later comments on that bug report,
which is not surprising because the change was wrong: this code path was
(and still is) used only for changing locations using the pathbar and
sidebar. The bug happening with double-click on the view is uses another
code path, in mime-actions.c, which has later been patched[2].

So, let's remove the wrong FIXME comment and effectively revert [0],
which, as expected, doesn't introduce any bug in my testing.

[0] 49c03251ee4b1de798e5675d74a7d30d281f22e8
[1] https://bugzilla.gnome.org/show_bug.cgi?id=756499
[2] 755c7710582eeb1cb34350ce983a6c4e3a451f47