files-view: Disable zoom on empty views
We have a visibly outdated comment on `real_end_file_changes()`, which
still talks about the long-gone zoom slider. But most importantly, it's
wrong: both zoom actions and sort actions are enabled on empty views
nowadays.

Yet, based on this outdated assumption, we proceed to compute the state
of every file action, which is expensive on very large directories.

As it turns out, the only single action which cares about whether the
view is empty nowadays is "select-all". So, let's avoid the expensive
call and update the enabled status of "select-all" alone.

While we are at it, also disable the zoom actions on empty views, as
there really is no visible effect and it's a simple change to make.

As for the sort action, it would be too troublesome to do the same thing
as that action is too entangled with NautilusListBase's properties and
there is not much benefit, because the result of changing sort option is
visible in the sort menu even without files in view.