window: Import empty-trash action
The action definition is duplicated between sidebar and the view. And
in the view we trigger expensive action state computations every time
the trash state changes, which is unnecessary just to update one single
action.

The upcoming FileChooser window is not going to feature trash nor an
empty-trash action, so make it a NautilusWindow action.