files-view: Always delay menu updates if inactive
In many places we avoid calling `schedule_update_context_menus()` if
the view is not in the active slot (i.e., current tab in the window).

This is a neat performance optimization for global events.

Let's use it in every case. Also use a flag to check whether we
actually need to update context menus when going active. And drop a
redundact `update_context_menus()` call at the same time.