files-view: Drop a couple unnecessary menu updates
When toggling hidden files, we reload the directory, which already
causes the context menus to be updated.

When copy or move operations are done, if they affect the current
view, we also update context menus from directory content change
signals. Also the comment is outdated, as undo/redo are not view
actions anymore, and they are updated elsewhere.