window-slot: Introduce :mode property
The slot and its internals are going to be reused by the upcoming
FileChooser window, as part of a FileChooser portal implementation.

However, some behaviors and UI elements are going to be different.
So, the slot needs to know on which mode to operate from the start.

There are 3 FileChooser portal methods[0]:

  - OpenFile
  - SaveFile
  - SaveFiles

But OpenFile has 2 boolean options ("directory" and "multiple) which
mean there are actually 4 modes:

  1. select single file
  2. select single folder
  3. select multiple files
  4. select multiple folders

As such, we have 6 new modes to support, in addition to the "browse"
mode (i.e., business as usual).

[0]
https://flatpak.github.io/xdg-desktop-portal/docs/doc-org.freedesktop.portal.FileChooser.html