file-utilities: add scripts to context menu in flatpak
The g_get_user_data_dir() function fetches the data directory without
considering if Nautilus is running in a sandbox or not. This results in
an incorrect path under the flapak sandbox, making scripts inaccessible.
We fix this by checking whether nautilus is sandboxed and adjust the
path accordingly

Fixes #3477