window-slot: Handle file activation on FileChooser
The view handles activation through mime-actions.c functions, which
allow, among other things, opening files in other apps.

We don't want that in the FileChooser modes. Opening folders is okay,
but what we want to do when double-clicking a file depends on the
exact mode and the filters that apply.

Prepare to handle activation at a higher lever: the expected action
will be implemented by the upcoming NautilusFileChooser, by handling
the new NautilusWindowSlot::activate-files signal.