window-slot: Make :selection property reliable
Currently, it caches the view selection through bindings.

Unfortunately, this cache may get out of sync with the view, because
NautilusView:selection is a file list that's generated on-demand, and
NautilusView::notify:selection is unreliable. E.g. when the view
switches to a new location, it empties the model, which effectively
clears selection, but this is not notified as a selection change.

Making NautilusView:selection's notifications reliable is going to
require quite a bit of refactoring, so that's for another time.

Drop the binding and proxy the selection property without caching it.