window-slot: Don't extract from URI on FileChooser
When asked to open the URI leading to an archive file, the slot
marks the file for extraction as soon as the view is loaded.

In FileChooser modes we wouldn't extract anyway, since activation
is handled by the slot. But it would lead to other unexpected
results, like immediately sending the file to the portal as result.

This is really a corner case but let's avoid it anyway.