mime-actions: Open locations directly on slot
While the FileChooser will not want to launch files in handler apps
on activation, it still needs to activate items that open in view.

For regular folders it would be trivial, but for mountables it would
have to reimplement logic that's already present in mime-actions.c

In order to make it possible to use `nautilus_mime_activate_files()`
from the FileChooser to open items in view, we need to stop relying
a NautilusApplication method, because the upcoming FileChooser will
not be a NautilusApplication window.

Since we already pass the slot as a pointer, we can just use the slot
method directly, which produces the same effect here anyway.