toolbar: Drop separators
They are hidden on narrow browser windows. And wide windows end up
having whitespace when the pathbar reaches maximum width.

But on FileChooser they are always visible and cause the toolbar not
to fit the minimum width on narrow windows.

Let's just remove the separators along with the boxes packing them
with the history and view controls.