portal: Handle SaveFile method
Saving a file includes choosing a filename.

GTK's file chooser features a filename entry which is always visible.
This entry is focused by default, which means typing will change the
name by default. This is a problem because in all other cases typing
filters the view instead (i.e. starts a search). But once you learn
this quirk, you get used to it... and that leads to another problem.

Let's suppose you want to pick a destination folder before deciding
on the filename (because you may want to look at existing filenames
on that folder first). Since you have changed locations, the focus is
now on the view, not the filename entry. (Yet, the filename is still
selected in the entry, for extra misleads.) If you've taught yourself
that typing will change the filename, surprise, it searches now!

For the nautilus implementation we shall avoid these pitfalls. Let's
do so with a slightly different solution: by default, the filename
widget is a button, which turns into an entry only if activated, and
reverts back to a button it loses focus. This way, we have consistent
type-to-search for all file chooser modes, yet we still preserve the
ability to change the filename while being able to see the content of
the destination folder.

We also introduce a "revert" button to restore the original filename
as suggested by the client application.

Part of: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3401