files-view: Grab focus correctly on new window
We have been carrying a gtk_widget_grab_focus() call to ensure the
focus is where we expect it to be when we begin loading a view. If
the search bar is not open, we expect the focus to be on the view.

However, grabbing focus on the view works only if it is parented.
This is not the case when loading the view for the first time,
because the slot doesn't parent the view until later.

As the grab_focus() call fails, the focus on new windows falls on the
first effectively focusable child of the toolbar (itself the first
child of the main pane). This default focus widget used to be the
first pathbar button, because the back/forward buttons were disabled.

But this has changed due to the back/forward actions being moved into
the slot.[0] Now, the default focus widget is the Back button. Then
this button becomes insensitive when the location is set on the slot
and the slot.back action is disabled. Thus, focus is lost, i.e., the
window's :focus-widget becomes NULL.

As a result, no keyboard shortcut managed by NautilusShortcutManager
works on newly opened windows. Only the app-global shortcuts work.
In particular, this means the tilde (`~`) keyboard shortcut doesn't
trigger but instead starts a search.

Instead of listening to notification on the NautilusView:loading
property, which is notified very early on, let's ensure the focus at
a later point: the NautilusFilesView::begin-loading signal.

Indeed, ::begin-loading is completely unrelated to :loading. The
later is meant refers to loading the view itself (which may start
before it's parented), while the former is about loading the contents
of the viewed directory, which happens at a later point (when the
view is already parented).

[0] commit 2d6f4b8280bdf58421194f0d94fff1f45f4a54c5