window: Drop unnecessary key capture controller
Originally introduced in preparation for the GTK4 port[0] to carry on
an earlier bugfix[1] back from early GTK3 years.

But it doesn't seem to be needed anymore. Possibly it was never
needed in GTK4 due to how it unified keyboard shortcuts.

Drop it. Both the case in the code (tilde) and the case in bugzilla
(Ctrl+A) work as expected, being handled by the entry instead of
triggering a our shortcut actions.

[0] commit 2d675e6c1a2bd06a3ba596bbd5a4b01b798cf025
[1] commit 8d3289d3f5386fac9fb2aded7dd67fd0a9252dd8