gtkplacessidebar: Open volume in slot that was active at start of mount
If the user change window slots between the time g_volume_mount started and completed, we always mount on the current active slot which may be disruptive to the user workflow.

This is because we use the sidebar's current window slot reference which is bound to the NautilusWindow active-slot property.

Keep a weak reference to the slot at the start of mount and open the volume there once mounted. If the slot has been disposed then use the current active slot.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2855