window-slot: Assert slot type after cancellation check
The NAUTILUS_WINDOW_SLOT assertion can fail if the slot is disposed and finalized after the start of g_file_find_enclosing_mount_async.

Although we cancel in dispose, GIO conventions mean it will complete the operation asynchronously and call our callback.

Move the assertion to after the cancellation check.