list-base: Always use the top-left corner for drag
This drops the custom dragging hotspot for the grid view, as it doesn't
align well with the clicked area, is inconsistent with the list and
network views, and doesn't make much sense when dragging multiple files.

This also avoids tapping into the internals of the list widgets only to
do speculative downcasts to have operations specific to each list
widget.

Signed-off-by: Adrien Plazas <aplazas@gnome.org>