touch: Change long-press propagation phase
The GtkGestureLongPress that's responsible for opening
up the background context menu was not receiving any
motion events, presumably due to the GtkGridView catching
them in the capture-phase already, and was thus unable
to detect passing the touch-drag threshold.
To re-enable items to be long-pressed as well, the background
long-press callback checks if the pointer is above an item
beforehand.

Fixes #2872