window-slot: Update back forward actions upon restore
After commit [1], back and forward actions are not correctly
reenabled after restoring a tab.  The actions were previously
reenabled after restoring a tab due to the fact that the
restored tab was not automatically active.  Once the tab
was created and the user switched to the newly restored tab
that would fire a location change which in turn called
nautilus_window_sync_location_widgets() where the back and
forward actions would get enabled.  Now that the actions live
in the slot, this no longer happens.  Reenable the actions
after the back and forward lists have been restored.

[1] 2d6f4b8280bdf58421194f0d94fff1f45f4a54c5