files-view: Make ::notify::selection  reliable
The changes notifications on the NautilusView:selection property are
not reliable. E.g. when the view switches to a new location, it
empties the model, which effectively clears selection, but this is
not notified as a selection change.

This happens because we have been incorrectly relying exclusively on
the ::selection-changed signal from the view model. As documented on
GtkSelectionModel, and contrary to intuition, this signal  doesn't
notify about changes to the selection set; instead, it notifies about
changes to the selection state of individual items! In other words,
it doesn't notify us if a selected item is removed from the model.

The FileChooser dialog is going to rely on NautilusWindowSlot to get
the current selection. The slot currently caches the selection from
the view. So, we need to invalidate that cache correctly, and for
this we need the view to reliably notify any changes to selection.

The correct way to effectively track changes to the selection set is
to listen to ::items-changed in addition to ::selection-changed.
Instead of reinventing the wheel, rely on GtkSelectionFilterModel to
do the job for us. Then we can just listen to changes in that model.

As it turns out, this is also the correct fix for
https://gitlab.gnome.org/GNOME/nautilus/-/issues/2338 and
https://gitlab.gnome.org/GNOME/nautilus/-/issues/3354

In the future we may want to further refactor our get_selection()
logic around this filter list model, perhaps with GtkMapListModel to
have NautilusFile as item-type, to use proper reference-counted
contained instead of going around doing deep copies of GLists, but
that's for another time.