window-slot: Fix implementation of :selection
It's missing a case in get_property() despite being readable. We
haven't been reading it so far, which is why this has gone unnoticed.
But the upcoming FileChooser implementation will read it.

Also, we have been explicitly notifying it, so flag it as such to
avoid double notification. While here, drop obsolete nick and blurb.