filename-validator: Ensure consistent state on notification
I've hit a problem where I wanted to check for the value of one
property when another one changes. But the former's value was
still not updated, because it's updated just after the later is
notified.

Since all of these properties deal with the validation of the
filename, they are expected to be updated at once. Therefore,
bundle the notifications at the end of the processing.