compress-dialog: Check selected item in bind
The checkmark is not showing up initially because we are binding the combo row's selected-item property in setup, when the list view has not yet been rooted. Because of this, we miss the notification from the initial setting of the item (the list item still doesn't have an item associated with it). By the time the list view is rooted, there are no further changes made.

Instead of binding the property in setup, listen for changes in bind at which point the item has been associated with the list item.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3523