new-folder-dialog: Actually modernize it
As a follow-up to
https://gitlab.gnome.org/GNOME/nautilus/-/merge_requests/1395

Use the new forming design pattern for action dialogs with a close
button and the suggested action centered at the bottom.

In order not to move the button at the bottom, the revealer no longer
uses a slide animation, but a crossefade instead, which means the
dialog content has a fixed height. As such, there is no need to
hardcode a dialog height to fit the releaver content anymore.