file-chooser: Adjust the file type width
Let it be a little longer, but make sure it doesn't completely disappear.
I'm not sure middle is the right way to ellipsize either, it seems to always
be wrong, whereas end is wrong sometimes (when the abbreviation is at the
end "Portable Document Format (PDF)"), and start is wrong when it isn't
"SVG Images".