pathbar: Use scroll-to api
Commit 6a0f483265 ("pathbar: Fix automatically scrolling") added
a 800ms animation in order to scroll in the next main context iteration.
This was necessary because the scrolled window actually expands async
after the adjustment size has changed.  Since the original solution
was implemented, the scroll-to api was introduced to GtkViewPort.
Remove the workaround and use the new api to keep the current folder
scrolled into view, without the visual defects added by the original
workaround.