pathbar: Fix automatic scrolling
Setting the adjustment value immediately on page size chage
creates a bug, where, when new windows are launched (via browse
or file chooser), the scroll is still set to the beginning (and the
current folder is out of view). This causes the same issues that
Commit 6a0f483265 ("pathbar: Fix automatically scrolling") fixed.
Unfortunately that commit does not address the issue because the
adjustment is already at the upper value. Remove this hack and
reintroduce the visual defect, which will be resolved in the next
commit.