file-chooser: Don't nest .toolbar styles
The pill button needs more breathing space than what's provided by
default for buttons in toolbars.

This is currently achieved by having .toolbar style in a child of a
.toolbar style container, resulting in double padding and spacing.

But this is obviously a hack and the double spacing is uncalled for,
affecting all other toolbar children and not only the pill button.

Let's instead add margin to the pill button itself.

The undo button is no longer inside a toolbar-styled container, so
make it explicitly flat instead.