file-chooser: Use a bin as slot container
It's awkward for to use an AdwToolbarView for this. Looking only
at the XML, it's not clear at all how the slot fits.

Instead, let's declare a bin in the XML as the content of the toolbar
view, then pack the slot as its child in code.

This also gives us flexibility to make future changes in layout.